chore(deps): Add Dependabot configuration for npm and GitHub Actions
This commit adds a `.github/dependabot.yml` file to enable automated dependency updates. Dependabot is configured to update `npm` and `GitHub Actions` dependencies on a weekly schedule. It also limits the number of open pull requests to 5 and specifies commit message prefixes as "deps" for npm and "ci" for GitHub Actions. This setup ensures consistent and timely updates, improving project maintenance and security.

Author: Noah Leibetseder